\name{datasim}
\alias{datasim}
\title{Datos simulados}
\description{
 simulated MTAR process
}
\usage{data(datasim)}
\format{Object with class \code{mtarsim}}
\source{R Simulate}
\references{
  Calderon V, S. A., & Nieto, F. H. (2017).  \emph{Bayesian analysis of multivariate       threshold autoregressive models with missing data}. Communications in Statistics-Theory   and Methods, 46(1), 296-318.
}
\keyword{datasets}
