\name{datasim_numreg}
\alias{datasim_numreg}
\title{Multivariate threshold autoregressive process simulation for estimate number of regimes}
\description{
 object of class \dQuote{\code{regime_number}}
}
\usage{data(datasim_numreg)}
\format{Object with class \code{regime_number}}
\source{R Simulate}
\references{
  Calderon V, S. A., & Nieto, F. H. (2017).  \emph{Bayesian analysis of multivariate       threshold autoregressive models with missing data}. Communications in Statistics-Theory   and Methods, 46(1), 296-318.
}
\keyword{datasets}
