\name{dmnormB}
\alias{dmnormB}
\title{Multivariate normal density using Brobdingnag class}
\usage{
dmnormB(x,mean,sigma)
}
\arguments{
\item{x}{numeric type object, value to compute the density of multivariate normal distribution. Not NULL}
\item{mean}{numeric type, mean of multivariate normal distribution. Not NULL}
\item{sigma}{matrix type object, covariance parameter of multivariate normal distribution. Not NULL}
}
\value{
object class \code{Brobdingnag}
}
\description{
Compute multivariate normal density with class \code{Brobdingnag} objects.
}
