\name{dwishartB}
\alias{dwishartB}
\title{Wishart density using Brobdingnag class}
\usage{
dwishartB(x,nu,S)
}
\arguments{
\item{x}{matrix type object, value to compute the density of Wishart distribution. Not NULL}
\item{nu}{numeric type, degrees of freedom. Not NULL}
\item{S}{matrix type object, parameter of Wishart distribution. Not NULL}
}
\value{
object class \code{Brobdingnag}
}
\description{
Compute Wishart density with class \code{Brobdingnag} objects.
}
