\name{hydrodata}
\alias{hydrodata}
\title{Hydrological data of Colombia}
\description{
Diary rainfall (in mm) and the diary river flow (in m^3/s) of
two rivers where a river empties into the other one in a region of department of Cauca in
Colombia.
}
\usage{data(hydrodata)}
\format{data.frame}
\source{IDEAM, the oficial Colombian agency for hydrological and meteorological studies}
\references{
  Calderon V, S. A., & Nieto, F. H. (2017).  \emph{Bayesian analysis of multivariate       threshold autoregressive models with missing data}. Communications in Statistics-Theory and Methods, 46(1), 296-318.
}
\keyword{datasets}
