\name{missingest}
\alias{missingest}
\title{Datos simulados}
\description{
results example missing data estimation process for a mtar process
}
\usage{data(missingest)}
\format{Object whit class \code{mtar_missing}}
\source{R Simulation}
\references{
  Calderon V, S. A., & Nieto, F. H. (2017).  \emph{Bayesian analysis of multivariate       threshold autoregressive models with missing data}. Communications in Statistics-Theory   and Methods, 46(1), 296-318.
}
\keyword{datasets}
