\name{mtarNAIC}
\alias{mtarNAIC}

\title{Compute NAIC of a MTAR model}
\description{
Compute NAIC of a ``regime_model'' class object
}
\usage{
mtarNAIC(regimemodel)
}

\arguments{
  \item{regimemodel}{object of class ``regime_model''}
}

\details{
  Detalles de los metodos usados en la funcion corespondiente
}
\value{
  Return a list type object
  $AICj numeric type, AIC for each regime
  $NAIC numeric type, NAIC value
}
\references{
  Fox, J. (2016)
  \emph{Applied Regression Analysis and Generalized Linear Models},
  Third Edition. Sage.
}
\author{Valeria Bejarano \email{vbejaranos@unal.edu.co} & Andrey Rincon \email{adrincont@unal.edu.co}}

\examples{
  poner ejemplo de uso de funcion con datos reales
}

\keyword{AIC}
\keyword{NAIC}
