\name{print}
\alias{print}
\title{print an object appropriate to a particular data type}
\usage{
print(object, ...)
}
\arguments{
\item{object}{an object, whose class will determine the behaviour of autoplot}

\item{...}{other arguments passed to specific methods}
}
\value{
print an output
}
\description{
\code{print} a particular output for an object of a
particular class in a single command. This defines the S3 generic that
other classes and packages can extend.
}
