\name{prodB}
\alias{prodB}
\title{Function to make product of elements of a list}
\usage{
prodB(x)
}
\arguments{
\item{x}{list of objects to make the product}
\item{...}{other arguments passed to specific methods}
}
\value{
float number result of the product
}
\description{
Auxiliary function of some functions in the MTAR package
}
