\name{datasim}
\alias{datasim}
\title{Datos simulados}
\description{
  Contingency Table
}
\usage{data(datasim)}
\format{Object whit class \code{mtar.sim}}
\source{DAPD (1997), Population, stratification and socioeconomic aspects of Bogota}
\references{C.E. Pardo y J.E. Ortiz (2004)
}
\keyword{datasets}
