\name{datasim}
\alias{datasim}
\title{Simulated Multivariate threshold autoregressive process simulation}
\description{
 simulated MTAR process
}
\usage{data(datasim)}
\format{Object with class \code{mtarsim}}
\source{R Simulate}
\references{
Calderon, S. and Nieto, F. (2017) \emph{Bayesian analysis of multivariate threshold autoregress models with missing data.} Communications in Statistics - Theory and Methods 46 (1):296–318. doi:10.1080/03610926.2014.990758.
}
\keyword{datasets}
