\name{datasim_numreg}
\alias{datasim_numreg}
\title{Multivariate threshold autoregressive process simulation for estimate number of regimes}
\description{
 object of class \dQuote{\code{regime_number}}
}
\usage{data(datasim_numreg)}
\format{Object with class \code{regime_number}}
\source{R Simulate}
\references{
Calderon, S. and Nieto, F. (2017) \emph{Bayesian analysis of multivariate threshold autoregress models with missing data.} Communications in Statistics - Theory and Methods 46 (1):296–318. doi:10.1080/03610926.2014.990758.
}
\keyword{datasets}
