\name{missingest}
\alias{missingest}
\title{simulated data}
\description{
results example missing data estimation process for a mtar process
}
\usage{data(missingest)}
\format{Object whit class \code{mtar_missing}}
\source{R Simulation}
\references{
Calderon, S. and Nieto, F. (2017) \emph{Bayesian analysis of multivariate threshold autoregress models with missing data.} Communications in Statistics - Theory and Methods 46 (1):296–318. doi:
10.1080/03610926.2014.990758.
}
\keyword{datasets}
