\name{repM}
\alias{repM}
\title{Function to create list of matrix objects}
\usage{
repM(M,r)
}
\arguments{
\item{M}{matrix type object}
\item{r}{integer indicating the length of the list}
}
\value{
List of matrix of length r
}
\description{
Auxiliary function of some functions in the MTAR package
}
